refactor: MenuGroup 서비스 리팩터링
- 패키지 구조 변경
- MenuGroup 목록 조회 인수테스트 추가
