+++ src/main/java/kitchenpos/menu/dto/MenuGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.menu.dto;
 
 import kitchenpos.domain.MenuGroup;
