+++ src/test/java/kitchenpos/domain/MenuTest.java
@@ -1,0 +1,7 @@
+package kitchenpos.domain;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class MenuTest {
+
+}
