+++ src/test/java/kitchenpos/menu/acceptance/MenuGroupAcceptanceTest.java
@@ -1,0 +1,42 @@
+package kitchenpos.menu.acceptance;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import io.restassured.RestAssured;
+import io.restassured.http.ContentType;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import kitchenpos.AcceptanceTest;
+import kitchenpos.menu.dto.MenuGroupRequest;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
+
+@DisplayName("메뉴 그룹을 관리한다.")
+public class MenuGroupAcceptanceTest extends AcceptanceTest {
+
+    @Test
+    @DisplayName("메뉴 그룹을 생성한다.")
+    void createMenuGroup() {
+        MenuGroupRequest menuGroupRequest = new MenuGroupRequest("추천메뉴");
+
+        ExtractableResponse<Response> response = RestAssured.given().log().all()
+            .contentType(ContentType.JSON)
+            .body(menuGroupRequest)
+            .when().post("/api/menu-groups")
+            .then().log().all().extract();
+
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+    }
+
+    @Test
+    @DisplayName("메뉴 그룹 목록을 조회한다.")
+    void findAll() {
+        ExtractableResponse<Response> response = RestAssured.given().log().all()
+            .contentType(ContentType.JSON)
+            .when().get("/api/menu-groups")
+            .then().log().all().extract();
+
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+}
